Move Card into card.h
